requirement.txt
future==0.16.0
six==1.11.0
django-crispy-forms==1.7.2
httplib2==0.11.3
django-formtools==2.1
django-reversion==2.0.13
django-import-export

mysqlclient==1.3.12

pillow==5.1.0